backend/requirements/base-requirements.txt
beautifulsoup4==4.4.1
bs4==0.0.1
Django==1.10.2
django-crispy-forms==1.6.0
django-filter==0.13.0
django-markdownx==1.5
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
dropbox==6.2
gevent==1.1.1
greenlet==0.4.10
grequests==0.3.0
gunicorn==19.6.0
lxml==3.6.0
Markdown==2.6.6
Pillow==3.2.0
PyJWT==1.4.2
redis==2.10.5
requests==2.10.0
six==1.10.0
tqdm==4.8.4
urllib3==1.15.1

# Own libraries
moonreader_tools
PyYAML==3.12
